Moves to Qt5 and builds on vs2010. Uses the Qt implementations of sha256 and json.

+++ server/server/sources/api/ApiSessions.cpp
@@ -96,7 +96,7 @@ QStringList ApiSessions::getSessions(const QString &id_account, const QString &c
         // Get the sessions that match the account
         query.prepare(Database::instance()->getQuery("Sessions", "getSessions"));
         query.bindValue(":id_account", id_account);
-        query.bindValue(":ignore_account", id_account.isEmpty());
+        query.bindValue(":ignore_account", (int)id_account.isEmpty());
         if (Database::instance()->query(query, result))
             for (i = 0, s = result.size(); i < s; ++i)
                 sessions << result[i]["id"].toString();
